requirements.txt
# ChwiliTranslate Dependencies

# UI Framework
PyQt6>=6.5.0

# OCR Engine
easyocr>=1.7.0

# Async HTTP for Translation APIs
aiohttp>=3.9.0

# Testing
pytest>=7.4.0
hypothesis>=6.92.0

# Screen Capture
mss>=9.0.0
Pillow>=10.0.0

# Encryption for API keys
cryptography>=41.0.0
